# docs/swagger.yaml
basePath: /api
definitions:
...
  dtos.CreatePositionDTO:
    properties:
      department_id:
        type: string
      nome:
        maxLength: 100
        minLength: 2
        type: string
    required:
    - department_id
    - nome
    type: object
...
  dtos.PositionResponseDTO:
    properties:
      created_at:
        type: string
      department_id:
        type: string
      id:
        type: string
      nome:
        type: string
      updated_at:
        type: string
    type: object
  dtos.PositionResultDTO:
    properties:
      created_at:
        type: string
      department_id:
        type: string
      department_nome:
        type: string
      id:
        type: string
      nome:
        type: string
      updated_at:
        type: string
    type: object
...
  dtos.UpdatePositionDTO:
    properties:
      department_id:
        type: string
      nome:
        maxLength: 100
        minLength: 2
        type: string
    type: object
...
  utils.FieldError:
    properties:
      field:
        type: string
      label:
        type: string
      message:
        type: string
      tag:
        type: string
      value: {}
    type: object
  utils.Payload:
    properties:
      error:
        description: 'ex: validation_failed, conflict, internal'
        type: string
      fields:
        description: detalhes por campo
        items:
          $ref: '#/definitions/utils.FieldError'
        type: array
      message:
        description: resumo humanizado
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: suporte@empresa.com
    name: Equipa de Desenvolvimento
  description: API de Gestão de Recursos Humanos (Províncias, Municípios, Departamentos,
    etc.)
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://localhost:8080/terms
  title: HRMS API
  version: "1.0"
paths:
...
  /positions:
    get:
      description: Lista todas as posições cadastradas
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dtos.PositionResponseDTO'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.Payload'
      summary: Listar posições
      tags:
      - Posições
    post:
      consumes:
      - application/json
      description: Cria uma nova posição (cargo) associada a um departamento
      parameters:
      - description: Dados da posição
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/dtos.CreatePositionDTO'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dtos.PositionResponseDTO'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Payload'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/utils.Payload'
      summary: Criar nova posição
      tags:
      - Posições
  /positions/{id}:
    delete:
      description: Remove uma posição com base no ID
      parameters:
      - description: ID da posição
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Payload'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.Payload'
      summary: Remover posição
      tags:
      - Posições
    get:
      description: Retorna os dados de uma posição específica
      parameters:
      - description: ID da posição
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dtos.PositionResponseDTO'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Payload'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.Payload'
      summary: Buscar posição por ID
      tags:
      - Posições
    put:
      consumes:
      - application/json
      description: Atualiza os dados de uma posição existente
      parameters:
      - description: ID da posição
        in: path
        name: id
        required: true
        type: string
      - description: Campos atualizados
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/dtos.UpdatePositionDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dtos.PositionResponseDTO'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Payload'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.Payload'
      summary: Atualizar posição
      tags:
      - Posições
  /positions/search:
    get:
      consumes:
      - application/json
      description: Busca cargos/posições com filtros e paginação
      parameters:
      - description: Texto de busca (nome da posição)
        in: query
        name: search
        type: string
      - description: Filtro por nome do departamento
        in: query
        name: filter
        type: string
      - description: Número máximo de registros a retornar
        in: query
        name: limit
        type: integer
      - description: Número de registros a ignorar (para paginação)
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dtos.PositionResultDTO'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.Payload'
      summary: Buscar posições
      tags:
      - Posições
...
schemes:
- http
swagger: "2.0"